move PorablePdbReader to seperate file and nit changes
