+++ src/Microsoft.TestPlatform.ObjectModel/Navigation/PortableSymbolReader.cs
@@ -8,25 +8,42 @@ namespace Microsoft.VisualStudio.TestPlatform.ObjectModel.Navigation
     using System.IO;
     using System.Linq;
     using System.Reflection;
-    using System.Reflection.Metadata;
-    using System.Reflection.Metadata.Ecma335;
     using System.Runtime.Loader;
 
     using Microsoft.VisualStudio.TestPlatform.Utilities.Helpers;
 
     /// <summary>
-    /// The portable symbol reader.
+    ///     The portable symbol reader.
     /// </summary>
     internal class PortableSymbolReader : ISymbolReader
     {
         /// <summary>
-        /// Key in first dict is Type FullName
-        /// Key in second dict is method name
+        ///     Key in first dict is Type FullName
+        ///     Key in second dict is method name
         /// </summary>
-        private Dictionary<string, Dictionary<string, DiaNavigationData>> methodsNavigationDataForType = new Dictionary<string, Dictionary<string, DiaNavigationData>>();
+        private Dictionary<string, Dictionary<string, DiaNavigationData>> methodsNavigationDataForType =
+            new Dictionary<string, Dictionary<string, DiaNavigationData>>();
+
+        /// <summary>
+        /// The cache symbols.
+        /// </summary>
+        /// <param name="binaryPath">
+        /// The binary path.
+        /// </param>
+        /// <param name="searchPath">
+        /// The search path.
+        /// </param>
+        public void CacheSymbols(string binaryPath, string searchPath)
+        {
+            this.PopulateCacheForTypeAndMethodSymbols(binaryPath);
+        }
+
+        /// <summary>
+        /// The dispose.
+        /// </summary>
         public void Dispose()
         {
-            foreach (Dictionary<string, DiaNavigationData> methodsNavigationData in this.methodsNavigationDataForType.Values)
+            foreach (var methodsNavigationData in this.methodsNavigationDataForType.Values)
             {
                 methodsNavigationData.Clear();
             }
@@ -35,11 +52,18 @@ namespace Microsoft.VisualStudio.TestPlatform.ObjectModel.Navigation
             this.methodsNavigationDataForType = null;
         }
 
-        public void CacheSymbols(string binaryPath, string searchPath)
-        {
-            PopulateCacheForTypeAndMethodSymbols(binaryPath);
-        }
-
+        /// <summary>
+        /// The get navigation data.
+        /// </summary>
+        /// <param name="declaringTypeName">
+        /// The declaring type name.
+        /// </param>
+        /// <param name="methodName">
+        /// The method name.
+        /// </param>
+        /// <returns>
+        /// The <see cref="INavigationData"/>.
+        /// </returns>
         public INavigationData GetNavigationData(string declaringTypeName, string methodName)
         {
             INavigationData navigationData = null;
@@ -55,6 +79,12 @@ namespace Microsoft.VisualStudio.TestPlatform.ObjectModel.Navigation
             return navigationData;
         }
 
+        /// <summary>
+        /// The populate cache for type and method symbols.
+        /// </summary>
+        /// <param name="binaryPath">
+        /// The binary path.
+        /// </param>
         private void PopulateCacheForTypeAndMethodSymbols(string binaryPath)
         {
             try
@@ -63,8 +93,7 @@ namespace Microsoft.VisualStudio.TestPlatform.ObjectModel.Navigation
                 using (var pdbReader = new PortablePdbReader(new FileHelper().GetStream(pdbFilePath, FileMode.Open)))
                 {
                     // Load assembly
-                    Assembly asm = AssemblyLoadContext.Default.LoadFromAssemblyPath(binaryPath);
-
+                    var asm = AssemblyLoadContext.Default.LoadFromAssemblyPath(binaryPath);
 
                     // Get all types to dict, fullname as key
                     var typesDict = asm.GetTypes().ToDictionary(type => type.FullName);
@@ -79,16 +108,13 @@ namespace Microsoft.VisualStudio.TestPlatform.ObjectModel.Navigation
                         {
                             if (string.CompareOrdinal(methodEntry.Value.Module.FullyQualifiedName, binaryPath) != 0)
                             {
-                                // Get source info for methods only defined in given binaryPath
                                 continue;
                             }
 
-                            var sourceInfo = pdbReader.GetSourceInformation(methodEntry.Value);
-                            if (sourceInfo != null)
+                            var diaNavigationData = pdbReader.GetDiaNavigationData(methodEntry.Value);
+                            if (diaNavigationData != null)
                             {
-                                methodsNavigationData.Add(
-                                    methodEntry.Key,
-                                    new DiaNavigationData(sourceInfo.Filename, sourceInfo.startLineNumber, sourceInfo.endLineNumber));
+                                methodsNavigationData.Add(methodEntry.Key, diaNavigationData);
                             }
                             else
                             {
@@ -107,133 +133,6 @@ namespace Microsoft.VisualStudio.TestPlatform.ObjectModel.Navigation
                 this.Dispose();
                 throw;
             }
-        }
-    }
-
-    internal class PortablePdbReader : IDisposable
-    {
-        private MetadataReader reader;
-        private MetadataReaderProvider provider;
-
-        public PortablePdbReader(Stream stream)
-        {
-            if (!IsPortable(stream))
-            {
-                throw new Exception("Given stream is not portable stream");
-            }
-
-            this.Setup(MetadataReaderProvider.FromPortablePdbStream(stream));
-        }
-
-        private void Setup(MetadataReaderProvider provider)
-        {
-            this.provider = provider;
-            this.reader = provider.GetMetadataReader();
-        }
-
-        public SourceInformation GetSourceInformation(MethodInfo methodInfo)
-        {
-            if (methodInfo == null)
-            {
-                return null;
-            }
-
-            var handle = methodInfo.GetMethodDebugInformationHandle();
-
-            return GetSourceInformation(handle);
-        }
-
-        private SourceInformation GetSourceInformation(MethodDebugInformationHandle handle)
-        {
-            if (this.reader == null)
-            {
-                throw new ObjectDisposedException(nameof(PortablePdbReader));
-            }
-
-            SourceInformation sourceInformation = null;
-            try
-            {
-                var methodDebugDefinition = this.reader.GetMethodDebugInformation(handle);
-                var fileName = GetMethodFileName(methodDebugDefinition);
-                int startLineNumber, endLineNumber;
-                GetMethodStartAndEndLineNumber(methodDebugDefinition, out startLineNumber, out endLineNumber);
-
-                sourceInformation = new SourceInformation(fileName, startLineNumber, endLineNumber);
-            }
-            catch (BadImageFormatException)
-            {
-            }
-
-            return sourceInformation;
-        }
-
-        private static void GetMethodStartAndEndLineNumber(MethodDebugInformation methodDebugDefinition, out int startLineNumber, out int endLineNumber)
-        {
-            var stratPoint =
-                methodDebugDefinition.GetSequencePoints().OrderBy(s => s.StartLine).FirstOrDefault();
-            startLineNumber = stratPoint.StartLine;
-            var endPoint =
-                methodDebugDefinition.GetSequencePoints().OrderByDescending(s => s.StartLine).FirstOrDefault();
-            endLineNumber = endPoint.StartLine;
-        }
-
-        private string GetMethodFileName(MethodDebugInformation methodDebugDefinition)
-        {
-            var fileName = string.Empty;
-            if (!methodDebugDefinition.Document.IsNil)
-            {
-                var document = this.reader.GetDocument(methodDebugDefinition.Document);
-                fileName = this.reader.GetString(document.Name);
-            }
-
-            return fileName;
-        }
-
-        public void Dispose()
-        {
-            this.provider?.Dispose();
-            this.provider = null;
-            this.reader = null;
-        }
-
-        private static bool IsPortable(Stream stream)
-        {
-            bool result = stream.ReadByte() == 'B' && stream.ReadByte() == 'S' && stream.ReadByte() == 'J' && stream.ReadByte() == 'B';
-            stream.Position = 0;
-            return result;
-        }
-    }
-
-    internal class SourceInformation
-    {
-        public SourceInformation(string filename, int startLineNumber, int endLineNumber)
-        {
-            Filename = filename;
-            this.startLineNumber = startLineNumber;
-            this.endLineNumber = endLineNumber;
-        }
-
-        public string Filename { get; }
-
-        public int startLineNumber { get; }
-
-        public int endLineNumber { get; }
-    }
-
-    internal static class MetadataExtensions
-    {
-        private static PropertyInfo methodInfoMethodTokenProperty = typeof(MethodInfo).GetProperty("MetadataToken");
-
-        internal static int GetMethodToken(this MethodInfo methodInfo)
-        {
-            return (int)methodInfoMethodTokenProperty.GetValue(methodInfo);
-        }
-
-        internal static MethodDebugInformationHandle GetMethodDebugInformationHandle(this MethodInfo methodInfo)
-        {
-            var methodToken = methodInfo.GetMethodToken();
-            var handle = ((MethodDefinitionHandle)MetadataTokens.Handle(methodToken)).ToDebugInformationHandle();
-            return handle;
         }
     }
 #endif
